Properly hide the switch transition animation for initialization by removing the `.loading` class only after forcing a style recomputation.
Adding a call to `getBoundingClientRect()` forces the browser to recompute the styles (which makes the `input:checked` dependent styles update) in order to get up-to-date dimensions. By doing so, we guarantee that the updating of the checkbox value and the removal of the `.loading` class aren't batched together, which would result in the CSS transition being shown. Taken from [this list](https://gist.github.com/paulirish/5d52fb081b3570c81e3a) of JS functions that force CSS/layout recomputation.

Usually the 5ms `setTimeout` call is enough to guarantee the styles are recomputed and the animations aren't shown when opening the popup, but this more strictly enforces the ordering and will prevent the transition from being shown on older, slower computers.