Animate the collapsible ports list to match prior behavior
TODO Once Firefox supports `interpolate-size: allow-keywords` we can remove this commit's JS additions in `PopupUI.js` and replace the CSS variable with `auto`.
Or it might be better to keep the JS approach for backwards compatibility with older browsers (Tor follows an ESR branch and might take a while to get support, for example).