Remove unneeded double-checking of same-origin requests (Fixes #59)
The double-checking erroneously fails different-protocol (eg. `ws://`) requests and also errors when `requestDetails.originUrl` is `undefined`, which occurs when pasting a URL into the bar instead of clicking a link.
It's safe to remove since it's entirely redundant (I initially added it to see what cases caused `requestDetails.thirdParty` to be true yet `originUrl.origin` and `request.origin` to differ, whoops).